refactor: change savings struct for balance, interest, and tax to float32 to handle decimals

+++ models/savings.go
@@ -1,10 +1,10 @@
 package models
 
 type Savings struct {
-	Id            int `json:"id" gorm:"primarykey"`
-	UserID        int `json:"user_id"`
-	SavingsNumber int `json:"savings_number"`
-	Balance       int `json:"balance"`
-	Interest      int `json:"interest"`
-	Tax           int `json:"tax"`
+	Id            int     `json:"id" gorm:"primarykey"`
+	UserID        int     `json:"user_id"`
+	SavingsNumber int     `json:"savings_number"`
+	Balance       float32 `json:"balance"`
+	Interest      float32 `json:"interest"`
+	Tax           float32 `json:"tax"`
 }
